Fixed bug in pasting that caused segfaults
The user_data of the lambda was being prematurely unrefed because
Gtk.Clipboard doesn't own the parameter. When data was available, the
unallocated memory was not yet modified, so it behaved correctly. When
the data is unavailable, many calls, presumably the ones to Xlib, would
reallocate the memory, causing the call back to be entered with garbage
data.

The change converts the lambda to a method call, and the reference to
this will outlive the clipboard.