Fixed bad mouse selection behaviour
The way start and stop were computed was wrong since the foreach loops
always proceed from start to stop in the positive direction. This meant
that if the user selected along a negative axis, the program would
select the complement of the region the user actually selected. This
behaviour has been fixed by storing the original mouse location and
updating the {start,stop}_{weft,warp} more intelligently.